+++ org.hive2hive.core.test/src/org/hive2hive/core/test/process/files/NewVersionTest.java
@@ -180,22 +180,22 @@ public class NewVersionTest extends H2HJUnitTest {
 		IFileConfiguration limitingConfig = new IFileConfiguration() {
 
 			@Override
-			public int getMaxSizeAllVersions() {
-				return Integer.MAX_VALUE;
+			public long getMaxSizeAllVersions() {
+				return Long.MAX_VALUE;
 			}
 
 			@Override
-			public int getMaxNumOfVersions() {
+			public long getMaxNumOfVersions() {
 				return 1;
 			}
 
 			@Override
-			public int getMaxFileSize() {
-				return Integer.MAX_VALUE;
+			public long getMaxFileSize() {
+				return Long.MAX_VALUE;
 			}
 
 			@Override
-			public int getChunkSize() {
+			public long getChunkSize() {
 				return H2HConstants.DEFAULT_CHUNK_SIZE;
 			}
 		};
@@ -226,22 +226,22 @@ public class NewVersionTest extends H2HJUnitTest {
 		IFileConfiguration limitingConfig = new IFileConfiguration() {
 
 			@Override
-			public int getMaxSizeAllVersions() {
-				return (int) fileSize;
+			public long getMaxSizeAllVersions() {
+				return fileSize;
 			}
 
 			@Override
-			public int getMaxNumOfVersions() {
-				return Integer.MAX_VALUE;
+			public long getMaxNumOfVersions() {
+				return Long.MAX_VALUE;
 			}
 
 			@Override
-			public int getMaxFileSize() {
-				return Integer.MAX_VALUE;
+			public long getMaxFileSize() {
+				return Long.MAX_VALUE;
 			}
 
 			@Override
-			public int getChunkSize() {
+			public long getChunkSize() {
 				return H2HConstants.DEFAULT_CHUNK_SIZE;
 			}
 		};
